+++ src/dom/checker.js
@@ -1,0 +1,21 @@
+import { isNullOrUndefined } from '@datatype/index.js';
+
+/**
+ * Verifies that an object is an *Element*
+ * @param {Element} obj 
+ * @returns {boolean} Value indicating whether the object is an *Element*
+ * @memberof DOM
+ */
+export function isElement(obj) {
+    return isNullOrUndefined(obj) ? false : obj.nodeType === 1 && obj instanceof Element;
+}
+
+/**
+ * Verifies that an object is an *HTMLElement*
+ * @param {Element} obj 
+ * @returns {boolean} Value indicating whether the object is an *Element*
+ * @memberof DOM
+ */
+export function isHTMLElement(obj) {
+    return isNullOrUndefined(obj) ? false : obj.nodeType === 1 && obj instanceof HTMLElement;
+}