+++ src/dom/index.js
@@ -2,5 +2,6 @@
 
 export * from './dom-create.js';
 export * from './dom-manip.js';
+export * from './checker.js';
 export * from './element-class-manip.js';
 export * from './element-getter.js';