+++ CHANGELOG.md
@@ -6,6 +6,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0 (2019-01-01)
 
-### Version 0.9.11 (2019-08-22)
+### Version 0.9.13 (2019-08-27)
+
+#### Added
+
+- Create helpers can add children elements
+
+#### Changed
+
+- Switch component look
+
+### Version 0.9.12 (2019-08-22)
 
 - Library is ready to be used with NPM (node package manager)